requirements.txt
beautifulsoup4==4.13.5
python-dotenv==1.1.1
Requests==2.32.5
